tools/manager_flask/requirements.txt
Flask==2.3.3
ruamel.yaml==0.17.32
jsonschema==4.19.0
Werkzeug==2.3.7
